refactor: make fund results evidence helpers separately testable
The results page was too large to review safely, so pure formatting and evidence helpers now live behind focused tests before JSX and hook extraction begins.

Constraint: Behavior-preserving refactor only; no route, payload, schema, or API changes.

Confidence: medium

Scope-risk: narrow

Tested: npx vitest client fund-model-results helper and page tests

Not-tested: Full client suite
